requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
Pillow==9.5.0
django-cors-headers==3.14.0
djoser==2.2.0